Rename config.js to fix aspnet:Config sub-generator test
