+++ MvcController/index.js
@@ -1,7 +1,7 @@
 'use strict';
 var util = require('util');
 var ScriptBase = require('../script-base.js');
-var cfg = require('../config');
+var cfg = require('../configuration');
 var path = require('path');
 
 var NamedGenerator = module.exports = function NamedGenerator() {
