+++ script-base-basic.js
@@ -9,7 +9,7 @@ var Generator = module.exports = function Generator() {
   this.sourceRoot(path.join(__dirname, './templates/'));
 
   this.namespace = function() {
-    return require('./config').getNamespace(this.fs);
+    return require('./configuration').getNamespace(this.fs);
   }.bind(this);
 };
